+++ cmd/gaia/cmd/gaiad/main.go
@@ -6,6 +6,7 @@ import (
 	"github.com/spf13/cobra"
 
 	abci "github.com/tendermint/abci/types"
+	tmtypes "github.com/tendermint/tendermint/types"
 	"github.com/tendermint/tmlibs/cli"
 	dbm "github.com/tendermint/tmlibs/db"
 	"github.com/tendermint/tmlibs/log"
@@ -26,7 +27,7 @@ func main() {
 
 	server.AddCommands(ctx, cdc, rootCmd, app.GaiaAppInit(),
 		server.ConstructAppCreator(newApp, "gaia"),
-		server.ConstructAppExporter(exportAppState, "gaia"))
+		server.ConstructAppExporter(exportAppStateAndTMValidators, "gaia"))
 
 	// prepare and add flags
 	executor := cli.PrepareBaseCmd(rootCmd, "GA", app.DefaultNodeHome)
@@ -37,7 +38,7 @@ func newApp(logger log.Logger, db dbm.DB) abci.Application {
 	return app.NewGaiaApp(logger, db)
 }
 
-func exportAppState(logger log.Logger, db dbm.DB) (json.RawMessage, error) {
+func exportAppStateAndTMValidators(logger log.Logger, db dbm.DB) (json.RawMessage, []tmtypes.GenesisValidator, error) {
 	gapp := app.NewGaiaApp(logger, db)
-	return gapp.ExportAppStateJSON()
+	return gapp.ExportAppStateAndValidators()
 }