+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 BREAKING CHANGES
 
+FEATURES
+
+IMPROVEMENTS
+* export command now writes current validator set for Tendermint
+
+FIXES
 * [lcd] Switch to bech32 for addresses on all human readable inputs and outputs
 
 ## 0.18.0